Add another function to return the spiking model instead of run_pipeline
in order to ensure compatibility even in the rare cases when GUI
is used.
Revert commit 279c054385e55d01793887695e604282c01fea3e,
and delegate all the pipeline down to the stage of creating the spiking
network to a separate function, and call it in run_pipeline().
That way, this function can be used to sidestep run_pipeline().